pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Chao Zhang <chao.zhang@midl.dev>"]
description = "Project for MIDL public status page"
name = "midl-status"
packages = [{include = "midl_status"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
celery = "^5.3.1"
flask-migrate = "^4.0.4"
flask-sqlalchemy = "^3.0.5"
gunicorn = "^21.2.0"
loki-api-client = "^0.1.0"
markupsafe = "^2.1.3"
psycopg2-binary = "^2.9.6"
python = "^3.10"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.0"
pytest = "^7.3.2"
pytest-celery = "^0.0.0"
pytest-flask = "^1.2.0"
toml-sort = "^0.23.1"
pytest-cov = "^4.1.0"
